Lexical: Reorganised custom node code into lexical codebase
Also cleaned up old unused imports.
