+++ resources/js/wysiwyg/lexical/core/__tests__/unit/LexicalEditor.test.ts
@@ -47,7 +47,6 @@ import {
 import invariant from 'lexical/shared/invariant';
 
 import {
-  $createTestDecoratorNode,
   $createTestElementNode,
   $createTestInlineElementNode,
   createTestEditor,
