+++ resources/js/wysiwyg/lexical/core/nodes/LexicalElementNode.ts
@@ -19,8 +19,8 @@ import invariant from 'lexical/shared/invariant';
 import {$isTextNode, TextNode} from '../index';
 import {
   DOUBLE_LINE_BREAK,
-  ELEMENT_FORMAT_TO_TYPE,
-  ELEMENT_TYPE_TO_FORMAT,
+
+
 } from '../LexicalConstants';
 import {LexicalNode} from '../LexicalNode';
 import {
