+++ resources/js/wysiwyg/lexical/readme.md
@@ -9,4 +9,4 @@ Only components used, or intended to be used, were copied in at this point.
 The original work built upon in this directory and below is under the copyright of Meta Platforms, Inc. and affiliates.
 The original license can be seen in the [ORIGINAL-LEXICAL-LICENSE](./ORIGINAL-LEXICAL-LICENSE) file.
 
-Files may have since been modified with modifications being under the license and copyright of the BookStack project as a whole. 
+Files may have since been added or modified with changes being under the license and copyright of the BookStack project as a whole.
