+++ resources/js/wysiwyg/lexical/rich-text/LexicalCodeBlockNode.ts
@@ -8,9 +8,9 @@ import {
     Spread
 } from "lexical";
 import type {EditorConfig} from "lexical/LexicalEditor";
-import {EditorDecoratorAdapter} from "../ui/framework/decorator";
-import {CodeEditor} from "../../components";
-import {el} from "../utils/dom";
+import {EditorDecoratorAdapter} from "../../ui/framework/decorator";
+import {CodeEditor} from "../../../components";
+import {el} from "../../utils/dom";
 
 export type SerializedCodeBlockNode = Spread<{
     language: string;
