+++ resources/js/wysiwyg/lexical/rich-text/LexicalHorizontalRuleNode.ts
