+++ resources/js/wysiwyg/lexical/rich-text/LexicalImageNode.ts
@@ -6,8 +6,8 @@ import {
     Spread
 } from "lexical";
 import type {EditorConfig} from "lexical/LexicalEditor";
-import {CommonBlockAlignment, extractAlignmentFromElement} from "./_common";
-import {$selectSingleNode} from "../utils/selection";
+import {CommonBlockAlignment, extractAlignmentFromElement} from "lexical/nodes/common";
+import {$selectSingleNode} from "../../utils/selection";
 import {SerializedElementNode} from "lexical/nodes/LexicalElementNode";
 
 export interface ImageNodeOptions {
