+++ resources/js/wysiwyg/ui/decorators/code-block.ts
@@ -1,7 +1,7 @@
 import {EditorDecorator} from "../framework/decorator";
 import {EditorUiContext} from "../framework/core";
-import {$openCodeEditorForNode, CodeBlockNode} from "../../nodes/code-block";
-import {$isDecoratorNode, BaseSelection} from "lexical";
+import {$openCodeEditorForNode, CodeBlockNode} from "@lexical/rich-text/LexicalCodeBlockNode";
+import {BaseSelection} from "lexical";
 import {$selectionContainsNode, $selectSingleNode} from "../../utils/selection";
